+++ frontend/app/src/screens/proposed-changes/item-info.tsx
